# docs/swagger.yaml
basePath: /
definitions:
  handlers.IDsRequest:
    properties:
      ids:
        items:
          type: integer
        type: array
    required:
    - ids
    type: object
  models.BulkTextCardsRequest:
    properties:
      cards:
        items:
          $ref: '#/definitions/models.TextCardInput'
        minItems: 1
        type: array
    required:
    - cards
    type: object
  models.CardFieldUpdate:
    properties:
      _id:
        type: integer
      count:
        type: integer
      id:
        type: integer
    type: object
  models.CardFull:
    properties:
      answer:
        items:
          $ref: '#/definitions/models.CardItemFull'
        type: array
      answerIds:
        items:
          type: integer
        type: array
      count:
        type: integer
      id:
        type: integer
      needed:
        type: integer
      parentNodes:
        items:
          type: integer
        type: array
      question:
        items:
          $ref: '#/definitions/models.CardItemFull'
        type: array
      questionIds:
        items:
          type: integer
        type: array
      reverseCount:
        type: integer
      shared:
        type: boolean
      usageType:
        $ref: '#/definitions/schema.UsageType'
      used:
        type: integer
      userId:
        type: integer
    type: object
  models.CardItemFull:
    properties:
      code:
        type: string
      extension:
        type: string
      formula:
        type: string
      id:
        type: integer
      imagePath:
        type: string
      index:
        type: integer
      shared:
        type: boolean
      text:
        type: string
      type:
        $ref: '#/definitions/schema.CardItemType'
      userId:
        type: integer
      width:
        type: string
    type: object
...
  models.CardItemShort:
    properties:
      code:
        type: string
      extension:
        type: string
      formula:
        type: string
      imagePath:
        type: string
      index:
        type: integer
      shared:
        type: boolean
      text:
        type: string
      type:
        $ref: '#/definitions/schema.CardItemType'
      width:
        type: string
    required:
    - type
    type: object
  models.CardPartial:
    properties:
      answer:
        items:
          $ref: '#/definitions/models.CardItemShort'
        type: array
      answerIds:
        items:
          type: integer
        type: array
      count:
        type: integer
      id:
        type: integer
      needed:
        type: integer
      parentNodes:
        items:
          type: integer
        type: array
      question:
        items:
          $ref: '#/definitions/models.CardItemShort'
        type: array
      questionIds:
        items:
          type: integer
        type: array
      reverseCount:
        type: integer
      shared:
        type: boolean
      usageType:
        $ref: '#/definitions/schema.UsageType'
      used:
        type: integer
    required:
    - id
    type: object
  models.CardsByQueryRequest:
    properties:
      group:
        $ref: '#/definitions/schema.CardsGroup'
      id:
        type: integer
      priority:
        $ref: '#/definitions/schema.CardsPriority'
      query:
        type: string
    required:
    - id
    type: object
...
  models.NewCardRequest:
    properties:
      _id:
        type: integer
      answer:
        items:
          $ref: '#/definitions/models.CardItemShort'
        type: array
      question:
        items:
          $ref: '#/definitions/models.CardItemShort'
        type: array
      shared:
        type: boolean
    required:
    - _id
    - answer
    - question
    type: object
...
  models.TextCardInput:
    properties:
      answerTextItems:
        items:
          type: string
        minItems: 1
        type: array
      questionTextItems:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - answerTextItems
    - questionTextItems
    type: object
  models.UpdateCardsFieldRequest:
    properties:
      cards:
        items:
          $ref: '#/definitions/models.CardFieldUpdate'
        type: array
      field:
        description: count
        type: string
    required:
    - cards
    - field
    type: object
...
  schema.CardItemType:
    enum:
    - TEXT
    - TEXT_WITH_HIGHLIGHTED_SYMBOL
    - CODE
    - FORMULA
    - IMAGE
    - WORD_WITH_STRESS
    - MARKDOWN
    type: string
    x-enum-varnames:
    - CardItemTypeText
    - CardItemTypeTextWithHighlightedSymbol
    - CardItemTypeCode
    - CardItemTypeFormula
    - CardItemTypeImage
    - CardItemTypeWordWithStress
    - CardItemTypeMarkdown
  schema.CardsGroup:
    properties:
      cards:
        items:
          type: integer
        type: array
      name:
        type: string
    type: object
  schema.CardsPriority:
    properties:
      cards:
        items:
          type: integer
        type: array
      name:
        type: string
      number:
        type: integer
    type: object
  schema.UsageType:
    enum:
    - active
    - passive
    - transitional
    - common
    type: string
    x-enum-varnames:
    - UsageTypeActive
    - UsageTypePassive
    - UsageTypeTransitional
    - UsageTypeCommon
...
host: localhost:3033
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    Memory nodes, cards, and card items API server.
    **Authentication:** Click **Authorize**, enter your email as username and your password, then Authorize. Swagger will call `POST /users/token` and attach the access token as a Bearer header.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: MemoryGuard API
  version: "1.0"
paths:
...
  /card/{id}:
    get:
      description: Returns a card with resolved question/answer card items for the
        authenticated user
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CardFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Get card by ID
      tags:
      - cards
  /cards:
    get:
      description: Returns cards for the given JSON ids query parameter owned by the
        authenticated user
      parameters:
      - description: JSON array of IDs, e.g. [1,2,3]
        in: query
        name: ids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CardFull'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: List cards by IDs (query)
      tags:
      - cards
  /cards-by-query:
    post:
      consumes:
      - application/json
      description: Returns cards for a memory node (or priority/group), filtered by
        a simple query string
      parameters:
      - description: Query request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CardsByQueryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CardFull'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Filter cards by query
      tags:
      - cards
  /decrease-card-count/{id}:
    put:
      description: Decrements the card count by 1 (not below 0)
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CardFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Decrease card count
      tags:
      - cards
  /delete-cards:
    post:
      consumes:
      - application/json
      description: Deletes owned cards by IDs (including shared cards owned by the
        user). For shared cards the user only has a grant to, removes the grant/count
        records instead of deleting the card.
      parameters:
      - description: List of card IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.IDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Delete cards
      tags:
      - cards
...
  /get-cards:
    post:
      consumes:
      - application/json
      description: Returns multiple cards by their IDs for the authenticated user
      parameters:
      - description: List of card IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.IDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CardFull'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Get cards by IDs
      tags:
      - cards
...
  /increase-card-count/{id}:
    put:
      description: Increments the card count by 1
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CardFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Increase card count
      tags:
      - cards
...
  /memory-node/{id}/cards:
    post:
      consumes:
      - application/json
      description: Creates multiple text-only cards under a memory node. Body may
        be a cards array or an object with cards. Each card requires questionTextItems
        and answerTextItems. Pass ?shared=true to mark all cards as shared.
      parameters:
      - description: Memory node ID
        in: path
        name: id
        required: true
        type: integer
      - description: Mark all cards as shared
        in: query
        name: shared
        type: boolean
      - description: Cards array or wrapper
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BulkTextCardsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CardFull'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Bulk create text cards
      tags:
      - cards
...
  /new-card:
    post:
      consumes:
      - application/json
      description: Creates card items and a card under a memory node for the authenticated
        user
      parameters:
      - description: New card request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.NewCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CardFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Create card
      tags:
      - cards
...
  /update-card:
    put:
      consumes:
      - application/json
      description: Partially updates an existing card owned by the authenticated user
      parameters:
      - description: Card update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CardPartial'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CardFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Update card
      tags:
      - cards
...
  /update-cards-field:
    post:
      consumes:
      - application/json
      description: Bulk-updates count on multiple cards owned by the authenticated
        user
      parameters:
      - description: Bulk field update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCardsFieldRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Bulk update card field
      tags:
      - cards
...
schemes:
- http
- https
security:
- Login:
  - api
securityDefinitions:
  Login:
    flow: password
    scopes:
      api: Access the MemoryGuard API
    tokenUrl: /users/token
    type: oauth2
swagger: "2.0"